% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCAplot.R
\name{BCAplot}
\alias{BCAplot}
\title{This is a ggplot function, which depends upon ggplot2 and ggpubr}
\usage{
BCAplot(x)
}
\arguments{
\item{x}{a dataframe of atleast two columns named "Conc" and "A595".}
}
\description{
ggpolot function to plot the summaried result of BCA standard curve (2nd degree polynomial model) as generated by prest().
}
\keyword{BCAplot,}
\keyword{prest}
