% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firepng.R
\name{firepng}
\alias{firepng}
\title{The firepng function}
\usage{
firepng(path, y, w = 6, h = 4)
}
\arguments{
\item{path}{file path in quotes.}

\item{y}{the R graphics object, which needs to be saved as a .svg file.}

\item{w}{width in inches, defaults to 6}

\item{h}{height in inches, defaults to 4}
}
\description{
Print a R Graphics object to a file.
}
\keyword{firepng}
