% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firesvg.R
\name{firesvg}
\alias{firesvg}
\title{The firesvg function}
\usage{
firesvg(path, y, w = 6, h = 6)
}
\arguments{
\item{path}{file path in quotes.}

\item{y}{the R graphics object, which needs to be saved as a .svg file.}

\item{w}{width in inches, defaults to 6}

\item{h}{height in inches, defaults to 6}
}
\description{
Print a R Graphics object to a file.
}
\keyword{firesvg}
