% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gelcast.R
\name{Gelcast}
\alias{Gelcast}
\title{Gelcast function for calculating required amounts of Lamelli SDS-PAGE components}
\usage{
Gelcast(runvol, runpc, stkvol, stkpc)
}
\arguments{
\item{runvol}{a numeric value representing the final volume of running gel mix required.}

\item{runpc}{a numeric value representing the running gel percentage required.}

\item{stkvol}{a numeric value representing the final volume of stacking gel mix required.}

\item{stkpc}{a numeric value representing the stacking gel percentage required.}
}
\description{
Lamelli SDS PAGE is cast using 5 components (30\% Acrylamide solution, Tris-HCl Buffer, Water, TEMED and 10\% APS). The Gelcast function calculates and prints the volumes of these components to be mixed, to cast a particular volume and percentage of running and stacking gels.
}
\keyword{Gelcast}
