% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prest.R
\name{prest}
\alias{prest}
\title{The prest function, which is short for protein estimation.}
\usage{
prest(x = 13, y)
}
\arguments{
\item{x}{row number in the dataframe where the unknown samples start, default is 13.}

\item{y}{row number in the dataframe where the unknown samples end.}
}
\description{
This function takes in a dataframe of "Sample", "Conc" and "A595".
The first 12 rows are for 6 standards in duplicates.
Depends on BCAplot(), summarySE().
}
\keyword{BCAplot,}
\keyword{prest,}
\keyword{summarySE}
\keyword{uroot,}
