% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uroot.R
\name{uroot}
\alias{uroot}
\title{The uroot function}
\usage{
uroot(x)
}
\arguments{
\item{x}{The numeric ordinate value (absorbance at 595 nm) for which the unknown numeric abcissa value (protein oncentration in mg/ml) is being sought.}
}
\description{
Calculates back x from a given y for a polynomial model defined as M1.
}
\keyword{uroot}
